Add a /status.js
This provides a javascript version of the same data in /status,
loading the information into a tiddlyweb.status variable.

This makes it possible to avoid an ajax call to determine things
like the current user name.

My original hope was to do this using content negotiation but
browsers in their infinite wisdom send Accept: */* when fulfilling
a <script> request.